Move and rename integration tests
